ContentHarvester class no longer makes sense
